% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNegative.R
\name{findNegative}
\alias{findNegative}
\title{Return the Negative Values in a Dataset}
\usage{
findNegative(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
the values in \code{x} that negative
}
\description{
Return the Negative Values in a Dataset
}
\examples{
findNegative(STaqiChange)
}
\keyword{find}
\keyword{negative,}
