% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPositive.R
\name{findPositive}
\alias{findPositive}
\title{Return the Positive Values in a Dataset}
\usage{
findPositive(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
the values in \code{x} that are positive
}
\description{
Return the Positive Values in a Dataset
}
\examples{
findPositive(STaqiChange)
}
\keyword{find}
\keyword{positive,}
