% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignificant.R
\name{findSignificant}
\alias{findSignificant}
\title{Find significant values above 30 or below -30 in a data frame}
\usage{
findSignificant(xb)
}
\arguments{
\item{xb}{A set of numbers, data frame or matrix}
}
\value{
True or false statements on whether \code{xb} data is significant
}
\description{
Find significant values above 30 or below -30 in a data frame
}
\examples{
findSignificant(STaqiChange)
}
\keyword{big,}
\keyword{large}
\keyword{number,}
