% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentChange.R
\name{pctC}
\alias{pctC}
\title{Percent change between two columns in a data frame}
\usage{
pctC(dataset, x, y)
}
\arguments{
\item{dataset}{Dataframe or matrix being used.}

\item{x}{Name of first column}

\item{y}{Name of second column}
}
\value{
The a column of percent changes from \code{x} to \code{y}
}
\description{
Percent change between two columns in a data frame
}
\examples{
pctC(temperature, temp80, temp18)
}
\keyword{change}
\keyword{percent,}
